HSEARCH-3511 Expect a (nullable) long as a parameter to setMaxResults/setFirstResult in the ORM mapper

+++ integrationtest/showcase/library/src/main/java/org/hibernate/search/integrationtest/showcase/library/repository/indexsearch/IndexSearchDocumentRepository.java
@@ -22,12 +22,12 @@ public interface IndexSearchDocumentRepository {
 
 	Optional<Book> getByIsbn(String isbnAsString);
 
-	List<Book> searchByMedium(String terms, BookMedium medium, int offset, int limit);
+	List<Book> searchByMedium(String terms, BookMedium medium, long offset, long limit);
 
 	List<Document<?>> searchAroundMe(String terms, String tags,
 			GeoPoint myLocation, Double maxDistanceInKilometers,
 			List<LibraryServiceOption> libraryServices,
-			int offset, int limit);
+			long offset, long limit);
 
 	List<String> getAuthorsOfBooksHavingTerms(String terms, SortOrder order);
 }
